backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-session==0.5.0
flask-limiter==3.5.0
SQLAlchemy==2.0.23
alembic==1.13.0
reportlab==4.0.7
duckdb==0.9.2
pandas==2.1.4
rapidfuzz==3.5.2
python-dotenv==1.0.0
waitress==2.1.2
cryptography==41.0.7
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
black==23.12.1
flake8==6.1.0
supabase==2.10.0
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
